Add Status Flags
It probably doesn't work...But we'll see what kind of amusing glitches it'll cause.

+++ NesCom/rom.cs
@@ -16,7 +16,7 @@ namespace NesCom
 	/// </summary>
 	public class rom
 	{
-		public int HeaderSize = 16;
+		public byte HeaderSize = 16;
 		int KBSize = 1024;
 		public byte NumOfCHRBlocks;
 		public byte NumOfPRGBlocks;
